【修改说明】 增加持久化模式选择 可选择file或者db方式存储

+++ spider-core/src/main/java/com/github/xdxiaodao/spider/core/service/biz/BookService.java
@@ -16,7 +16,7 @@ public interface BookService {
      * @param book 书籍信息
      * @return 新加书籍信息编号
      */
-    public int addBook(Book book);
+    public Long addBook(Book book);
 
     /**
      * 根据书名查询书籍信息
